fixed initial location on map load
