+++ src/components/Map/Map.jsx
@@ -14,8 +14,6 @@ export default function Map({
   showPopup,
   setShowPopup,
 }) {
-
-  
   return (
     <div className={styles.map}>
       <h1 className="text-slate-100">friends.map()</h1>
@@ -25,12 +23,11 @@ export default function Map({
         style={{ position: 'relative', width: '100%', height: '100%' }}
         mapStyle="mapbox://styles/mapbox/dark-v10"
         mapboxAccessToken={process.env.MapboxAccessToken}
-        onMove={e => setViewport(e.viewState)}
+        onMove={(e) => setViewport(e.viewState)}
       >
         <NavigationControl />
         <GeolocateControl
           positionOptions={{ enableHighAccuracy: true }}
-          trackUserLocation
           showUserHeading
           onGeolocate={(e) =>
             setUserCoords({
